+++ test/typescript/apis/project/dummy.ts
@@ -22,11 +22,18 @@ export class Bars extends Array<Bar> { static readonly drafts: typeof Foo }
 
 
 // for bound/ unbound actions
-export const action = ((foo: Foo) => 42) as Action
-export type Action = {
+export const unboundAction = ((foo: Foo) => 42) as UnboundAction
+export type UnboundAction = {
   (foo: Foo): number,
   __parameters: {foo: Foo},
   __returns: number
+}
+export const boundAction = ((foo: Foo) => 42) as BoundAction
+export type BoundAction = {
+  (foo: Foo): number,
+  __parameters: {foo: Foo},
+  __returns: number,
+  __self: Foo
 }
 
 // little 'trust me, type system!' helper to make T|undefined into T.