+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Added
 ### Changed
+- `req.subject` now points to the bound entity type when implementing handlers for bound actions.
 ### Deprecated
 ### Removed
 ### Fixed